wireguard: move IPCache to init param
This commit completes the set of changes needed so that every operation
on the Wireguard agent is registered from the agent itself upon starting
the cell. Here, we add IPCache as a dependency, and upon starting the agent,
in case of no errors, we autonomously register to all needed events.

Note: similarly for the IPSec key custodian, the WireGuard agent now
independently modifies the local node store upon starting. If we removed
the reference to the agent from the daemon params, the agent could
execute its Start routine after the LocalNode has already been written
to the CiliumNode/kvstore representation. This would cause problems,
given the agent (as for IPSec) sets the encryption key in the node object,
which is required for encryption to work (potential leakages observed
in node-to-node-encryption test otherwise).

we preserve the reference to the WireGuard agent in the daemon
params, so that we force it to start

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>